+++ src/arkanoid/FieldFactory.java
@@ -24,15 +24,15 @@ public class FieldFactory {
 		// Построение уровня
 		// TODO: Загрузка уровня из файла (пока уровень захардкоден)
 		ArkanoidField newField = new ArkanoidField(size);
-		BasicBall newball = new BasicBall(newField, new Point2D.Float(40, 160), 8, new Speed2D(0.03, -0.01));
-		BreakableBrick newbrick = new BreakableBrick(newField, new Point2D.Float(180, 120), new Dimension(48, 24));
-        BreakableBrick newbrick2 = new BreakableBrick(newField, new Point2D.Float(228, 120), new Dimension(48, 24));
-        UnbreakableBrick newbrick3 = new UnbreakableBrick(newField, new Point2D.Float(276, 120), new Dimension(48, 24));
-        BasicPaddle paddle = new BasicPaddle(newField, new Point2D.Float(0, 584), new Dimension(96, 16));
+		BasicBall newball = new BasicBall(newField, new Point2D.Double(40, 160), 8, new Speed2D(0.03, -0.01));
+		BreakableBrick newbrick = new BreakableBrick(newField, new Point2D.Double(180, 120), new Dimension(48, 24));
+        BreakableBrick newbrick2 = new BreakableBrick(newField, new Point2D.Double(228, 120), new Dimension(48, 24));
+        UnbreakableBrick newbrick3 = new UnbreakableBrick(newField, new Point2D.Double(276, 120), new Dimension(48, 24));
+        BasicPaddle paddle = new BasicPaddle(newField, new Point2D.Double(0, 584), new Dimension(96, 16));
         
         // Тестирование столкновения множества шаров
-        BasicBall ball01 = new BasicBall(newField, new Point2D.Float((float) 213.3975, 250), 16, new Speed2D(0.043, -0.025));
-        BasicBall ball02 = new BasicBall(newField, new Point2D.Float(400, 200), 16, new Speed2D(-0.05, 0));
+        BasicBall ball01 = new BasicBall(newField, new Point2D.Double((float) 213.3975, 250), 16, new Speed2D(0.043, -0.025));
+        BasicBall ball02 = new BasicBall(newField, new Point2D.Double(400, 200), 16, new Speed2D(-0.05, 0));
         ball01.addDefaultCollisionBehaviour(ReactionRebound.getInstance());
         ball02.addDefaultCollisionBehaviour(ReactionRebound.getInstance());
